Separate proxysql tests
